+++ Runtime/Basic/FloatVariable.cs
@@ -1,9 +1,10 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 
 namespace GI.UnityToolkit.Variables
 {
     [CreateAssetMenu(menuName = "Variable/Float")]
-    public class FloatVariable : DefaultedVariable<float>
+    public class FloatVariable : PersistableVariable<float>
     {
         #region PUBLIC_INTERFACE
 
@@ -23,5 +24,8 @@ namespace GI.UnityToolkit.Variables
         }
 
         #endregion PUBLIC_INTERFACE
+
+        protected override Action<string, float> GetSaveMethod() => PlayerPrefs.SetFloat;
+        protected override Func<string, float, float> GetLoadMethod() => PlayerPrefs.GetFloat;
     }
 }