+++ Runtime/Basic/StringVariable.cs
@@ -1,9 +1,10 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 
 namespace GI.UnityToolkit.Variables
 {
     [CreateAssetMenu(menuName = "Variable/String")]
-    public class StringVariable : DefaultedVariable<string>
+    public class StringVariable : PersistableVariable<string>
     {
         #region PUBLIC_INTERFACE
 
@@ -18,5 +19,9 @@ namespace GI.UnityToolkit.Variables
         }
 
         #endregion PUBLIC_INTERFACE
+
+        protected override Action<string, string> GetSaveMethod() => PlayerPrefs.SetString;
+
+        protected override Func<string, string, string> GetLoadMethod() => PlayerPrefs.GetString;
     }
 }