+++ Runtime/DataObject.cs
@@ -3,7 +3,7 @@
 namespace GI.UnityToolkit.Variables
 {
     /// <summary>
-    ///     Base class for all variables and states.
+    /// Base class for all variables and states.
     /// </summary>
     [CreateAssetMenu(menuName = "Variable/Empty")]
     public class DataObject : ManagedObject
@@ -26,10 +26,6 @@ namespace GI.UnityToolkit.Variables
 #endif
             // Don't want the object to die even if there are no references
             hideFlags |= HideFlags.DontUnloadUnusedAsset;
-        }
-
-        protected override void OnEnd()
-        {
         }
 
         #endregion MANAGED_OBJECT