+++ Runtime/PersistenceMode.cs
@@ -1,0 +1,9 @@
+namespace GI.UnityToolkit.Variables
+{
+    public enum PersistenceMode
+    {
+        None = 0,
+        ResetToDefault = 1,
+        SaveBetweenSessions = 2
+    }
+}