+++ Runtime/Variable.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using JetBrains.Annotations;
+using UnityEngine;
 using UnityEngine.Events;
 
 #if ODIN_INSPECTOR
@@ -10,26 +11,56 @@ using NaughtyAttributes;
 namespace GI.UnityToolkit.Variables
 {
     /// <summary>
-    ///     Base class of all shareable variable types.
+    /// Base class of all shareable variable types.
     /// </summary>
     /// <typeparam name="T"></typeparam>
-    public class Variable<T> : Variable
+    public abstract class Variable<T> : Variable
     {
         #region UNITY_EDITOR_INTERFACE
-
-        private void OnValueChanged()
-        {
-            if (Application.isPlaying == false) return;
-            OnChangedEvent?.Invoke();
-        }
-        
+        /// <summary>
+        /// The value of the variable. Used internally and in the editor.
+        /// </summary>
 #if ODIN_INSPECTOR
         [Title("Variable"), PropertyOrder(1), OnValueChanged(nameof(OnValueChanged))]
 #else
         [Header("Variable"), OnValueChanged(nameof(OnValueChanged))]
 #endif
-        [Tooltip("Current value this variable has in the game."), SerializeField] private T value = default;
+        [Tooltip("Current value this variable has in the game."), SerializeField]
+        private T value = default;
+
+        /// <summary>
+        /// A default value for the variable.
+        /// </summary>
+#if ODIN_INSPECTOR
+        [Title("Default"), PropertyOrder(2)]
+#else
+        [Header("Default")]
+#endif
+        [Tooltip("A default value for this variable. If the persistence method is \"Default\" then the variable will be set to this value on startup."), SerializeField]
+        private T defaultValue = default;
         
+        /// <summary>
+        /// How the value of the variable is set at startup.
+        ///     None
+        ///         - The value is not deliberately changed at startup.
+        /// 
+        ///     Reset To Default
+        ///         - The value is set to the specified default value at startup.
+        /// 
+        ///     Save Between Sessions
+        ///         - Changes to the value are saved between app sessions.
+        /// </summary>
+#if ODIN_INSPECTOR
+        [Title("Persistence"), PropertyOrder(2), ValueDropdown("AvailablePersistenceModes")]
+#else
+        [Header("Persistence"), Dropdown("AvailablePersistenceModes")]
+#endif
+        [Tooltip("How the value of this variable is handled at startup"), SerializeField]
+        protected PersistenceMode PersistenceMode = PersistenceMode.ResetToDefault;
+        
+        #endregion UNITY_EDITOR_INTERFACE
+
+        #region PUBLIC_INTERFACE
         public T Value
         {
             get => value;
@@ -37,14 +68,12 @@ namespace GI.UnityToolkit.Variables
             {
                 var oldValue = this.value;
                 this.value = value;
-                if (this.value == null && oldValue != null || this.value != null && !this.value.Equals(oldValue)) OnChangedEvent?.Invoke();
+                if (this.value == null && oldValue != null || this.value != null && !this.value.Equals(oldValue)) OnValueChanged();
             }
         }
-
-        #endregion UNITY_EDITOR_INTERFACE
-
-        #region PUBLIC_INTERFACE
-
+        
+        public T DefaultValue => defaultValue;
+        
         public void SetValue(T newValue)
         {
             Value = newValue;
@@ -61,20 +90,38 @@ namespace GI.UnityToolkit.Variables
         }
 
         /// <summary>
-        ///     Sets Value to the default value for the given type - "", 0, or null
+        /// Sets the variable to its specified default value.
         /// </summary>
-        public virtual void Default()
+        public void Default()
         {
-            Value = default;
+            var oldValue = Value;
+            Value = DefaultValue;
+            if (Value == null && oldValue != null || Value != null && !Value.Equals(oldValue)) OnValueChanged();
         }
-
         #endregion PUBLIC_INTERFACE
+        
+        #region OVERRIDES
+        protected override void OnBegin()
+        {
+            base.OnBegin();
+            if (PersistenceMode == PersistenceMode.ResetToDefault) Default();
+        }
+        #endregion
+        
+        [UsedImplicitly]
+        protected virtual PersistenceMode[] AvailablePersistenceModes => new[] { PersistenceMode.None, PersistenceMode.ResetToDefault };
+        
+        protected virtual void OnValueChanged()
+        {
+            if (Application.isPlaying == false) return;
+            OnChangedEvent?.Invoke();
+        }
     }
 
     public abstract class Variable : DataObject
     {
 #if ODIN_INSPECTOR
-        [Title("Events"), PropertyOrder(3)]
+        [Title("Events"), PropertyOrder(4)]
 #else
         [Header("Events")]
 #endif